+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.1...@esri/calcite-components-react@5.0.0-next.2) (2025-11-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.0...@esri/calcite-components-react@5.0.0-next.1) (2025-11-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.48...@esri/calcite-components-react@5.0.0-next.0) (2025-11-08)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.48](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.47...@esri/calcite-components-react@4.0.0-next.48) (2025-11-07)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.47](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.46...@esri/calcite-components-react@4.0.0-next.47) (2025-11-07)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [4.0.0-next.46](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.45...@esri/calcite-components-react@4.0.0-next.46) (2025-11-05)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [4.0.0-next.45](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@4.0.0-next.44...@esri/calcite-components-react@4.0.0-next.45) (2025-11-04)
 
 **Note:** Version bump only for package @esri/calcite-components-react